feat: node now dials out to connected peers
This change introduces a new event `PeerListManagerEvent::Dial(peer_id)` in the `PeerListManager` enum. When this event is triggered, the node will dial out to the specified peer. Additionally, a new field `dial_interval` has been added to the `PeerListManagerConfig` struct to configure the interval at which new peers will be dialed.

Changes:
- Added a new event `Dial(peer_id)` to the `PeerListManagerEvent` enum to trigger dialing a peer.
- Modified the `node.rs` file to handle the `Dial` event by dialing the specified peer.
- Added a new struct `DialInfo` to track dialing attempts and last dial time.
- Added enum `PeerState` to represent the state of a peer (Disconnected, Connected, Dialing).
- Modified `SimplePeerListManager` to include `PeerInfo` struct for peers, tracking their state and reputation.
- Added a method `connected_peers` to `SimplePeerListManager` to iterate over connected peers.